pyproject.toml
[tool.poetry]
name = "LostAndFound"
version = "0.1.0"
description = ""
authors = ["epe2 <nedellis1234@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
slackclient = "^2.5.0"
fastapi = "^0.47.1"
uvicorn = "^0.11.2"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[tool.poetry.scripts]
main = "lostandfound.core:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

